Resolve a few signedness mismatch warnings. (ml)
git-svn-id: svn://mielke.cc/main/brltty@5584 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/datafile.c
@@ -92,7 +92,7 @@ reportDataError (DataFile *file, char *format, ...) {
 }
 
 int
-isKeyword (const wchar_t *keyword, const wchar_t *characters, int length) {
+isKeyword (const wchar_t *keyword, const wchar_t *characters, size_t length) {
   if (length != wcslen(keyword)) return 0;
 
   while (length > 0) {
